WIP: Integrate visualizer into library
